requirements/base.txt
Django==5.0.7
python-environ==0.4.54
psycopg2-binary==2.9.9
django-ckeditor==6.7.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
pillow==10.4.0
django-mptt==0.16.0